udhcpc6: add ELAPSED_TIME option to outgoing packets
function                                             old     new   delta
init_d6_packet                                        53     121     +68
udhcpc_main                                         2577    2582      +5
udhcpc6_main                                        2593    2597      +4
------------------------------------------------------------------------------
(add/remove: 0/0 grow/shrink: 3/0 up/down: 77/0)               Total: 77 bytes

+++ networking/udhcp/dhcpc.h
@@ -22,8 +22,8 @@ struct client_data_t {
 	uint8_t *hostname;              /* Optional hostname to use */
 	uint8_t *fqdn;                  /* Optional fully qualified domain name to use */
 
-	uint16_t first_secs;
-	uint16_t last_secs;
+	unsigned first_secs;
+	unsigned last_secs;
 
 	int sockfd;
 	smallint listen_mode;
